+++ swing/common-model/src/test/java/is/codion/swing/common/model/worker/ProgressWorkerTest.java
@@ -63,7 +63,7 @@ public final class ProgressWorkerTest {
 						})
 						.onCancelled(() -> {})
 						.onInterrupted(() -> {})
-						.onException(throwable -> {})
+						.onException(exception -> {})
 						.execute()
 						.get();
 	}